requirements.txt
langchain==0.1.16
langchain-core==0.1.45
langchain-community==0.0.32
langchain-groq==0.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pymongo==4.6.3
python-dotenv==1.0.1